LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/responsive-iframe/responsive-iframe-value_value=content-inline-size,400,150-expected.txt


FAIL Measure target with value=content-inline-size and wm=undefined assert_array_equals: expected property 0 to be 400 but got 300 (expected array [400, 150] got [300, 150])
